+++ driver/cl_options.h
@@ -46,6 +46,7 @@ extern cl::list<std::string> runargs;
 extern cl::opt<bool> invokedByLDMD;
 extern cl::opt<bool> compileOnly;
 extern cl::opt<bool> emitDwarfDebugInfo;
+extern llvm::SmallVector<std::pair<std::string, std::string>, 0> debugPrefixMap; // Prefix map for filenames in DWARF debuginfo
 extern cl::opt<bool> noAsm;
 extern cl::opt<bool> dontWriteObj;
 extern cl::opt<std::string> objectFile;